requirements.txt
openai>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
